Simulation: add white flash on turn off
