+++ simulator/mocks/electrical/i_ic.h
@@ -8,6 +8,8 @@
 
 #include <cstdint>
 
+#include <memory>
+
 class IntegratedCircuitMock_I
 {
 public:
@@ -18,13 +20,60 @@ public:
   virtual uint8_t get_i2c_address() const = 0;
 
   // write data to this device
-  virtual int i2c_write_data(const uint8_t registerAddress, const uint8_t dataSize, const uint8_t* dataBuffer) = 0;
+  virtual int i2c_write_data(const uint8_t registerAddress, const uint8_t dataSize, const uint8_t* dataBuffer)
+  {
+    if (_registerMap[registerAddress] != nullptr)
+    {
+      uint16_t data = dataBuffer[0];
+      if (dataSize > 1)
+        data |= (dataBuffer[1] << 8);
+      return _registerMap[registerAddress]->write(data);
+    }
+
+    // failure
+    return 1;
+  }
+
   // read data from a register
-  virtual int i2c_read_data(const uint8_t registerAddress, const uint8_t dataSize, uint8_t* dataBuffer) = 0;
+  virtual int i2c_read_data(const uint8_t registerAddress, const uint8_t dataSize, uint8_t* dataBuffer)
+  {
+    if (_registerMap[registerAddress] != nullptr)
+    {
+      const uint16_t d = _registerMap[registerAddress]->read();
+
+      dataBuffer[0] = d & 0xff;
+      if (dataSize > 1)
+        dataBuffer[1] = (d >> 8) & 0xff;
+      return 0;
+    }
+
+    // failure
+    return 1;
+  }
+
   // simultaneous read write
-  virtual int i2c_xfer_data(const int outSize, const uint8_t* out, const int inSize, uint8_t* in) = 0;
+  virtual int i2c_xfer_data(const int outSize, const uint8_t* out, const int inSize, uint8_t* in)
+  {
+    // fail
+    return 1;
+  }
+
+  // register entry point
+  struct Register
+  {
+    uint16_t _data = 0;
+
+    virtual uint16_t read() { return _data; }
+    virtual int write(uint16_t data)
+    {
+      _data = data;
+      return 0;
+    }
+  };
 
 protected:
+  std::unique_ptr<Register> _registerMap[255];
+
 private:
 };
